Added a new field, donor ID, and tweaked AMR format
After a meeting with the consortium it was decided that we should add
another field, donor ID, a free text identifier for each sample that is
used internally by the depositing institute.

It was decided that the field "collected at" should be renamed as
"submitted by".

It was also agreed that the AMR strings should allow "U" (unknown) as a
valid SIR flag, and that MIC should be optional.

This commit updates the manifest template/notes and the checklist
configuration to reflect these changes.